A 3/4 implementation of networking between core and the dashboard
So here, the core manages to send the nparr of three channels over the
network as after encoding it as a string. The dashboard recieves the
data but since it does not know any information it about the array
except that it is composed of bytes, it cannot recreate it. What happens
exactly, the nparry is an array of arrays. It is an array of channel
arrays, so in the case of RBG it will be three arrays. Since the
dashboard does not know this information over the network and is reading
the data in chunks, it ends up constructing a new array with all the
elements in it, effectivly losing the channel information.
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
# On branch network
# Your branch is up-to-date with 'origin/network'.
#
# Changes to be committed:
#	modified:   core/core.py
#	modified:   dashboard/UI_Start.py
#	new file:   dump
#	new file:   printout
#