Finished the bare bone network implementation
For now, a pic gets sent as a string. The setup can tolerate changes in
image length and the client connecting and disconnecting. This does not
tolerate any packet drops, especially in the length sending phase.
Later, I will implement a check around that to make sure the connection
does not mess it up.

Also for now, each frame sends its length before it. I have yet to test
a uniform length for the entire connection period, but it seems that the
image sizes varry a lot for some reason.

In UI_Start.py after line 291 is where the pyQt parsing of whatever
image format I managed to send should be done. The image gets printed on
the CL if that helps, it is a single numpy array of bytes. Good luck
Will.
# On branch network
# Your branch is up-to-date with 'origin/network'.
#
# Changes to be committed:
#	modified:   dora/core/core.py
#	modified:   dora/dashboard/UI_Start.py
#